Fix QA issues and close test coverage gaps
Bug fixes (from qa-fixes plan):
- Fix telemetry counter reset-before-read in WidebandDecoder
- Fix encode_sint() signed overflow UB for INT64_MIN
- Fix channel energy double-counting in L1 FFT accumulation
- Pre-allocate and zero _gapBuf for overflow gap filling
- Pre-reserve HalfBandDecimator _workBuf per cascade stage
- Wrap all telemetry callbacks in try/catch (noexcept safety)
- Fix trx_perf.py dedup inflating frame_count
- Add socket.timeout catch for clean trx_perf.py exit
- Guard os_factor<1 with diagnostic log
- Fix parse_host_port rejecting port-only specs
- Remove dead peakBinFreq() and computeEnergySnapshot()

New test coverage:
- qa_lora_cbor: 28 tests for CBOR encode/decode + Telemetry encoder
- qa_lora_tx_queue: 4 tests for TxQueueSource burst scheduling
- test_trx_perf.py: 25 tests for frame stats and formatting
- test_lora_mon.py: +20 tests for sanitize_text, parse_host_port
- qa_lora_framesink: CBOR field-level decode verification
- qa_lora_burst: tightened CRC corruption test
- qa_lora_wideband: meaningful assertion for strong-tone test

+++ blocks/include/gnuradio-4.0/lora/cbor.hpp
@@ -56,7 +56,8 @@ inline void encode_sint(std::vector<uint8_t>& buf, int64_t val) {
     if (val >= 0) {
         encode_head(buf, 0, static_cast<uint64_t>(val));
     } else {
-        encode_head(buf, 1, static_cast<uint64_t>(-1 - val));  // CBOR major 1
+        // Use unsigned arithmetic to avoid UB when val == INT64_MIN
+        encode_head(buf, 1, static_cast<uint64_t>(-1) - static_cast<uint64_t>(val));
     }
 }
